test: (#65) BaseServiceTest를 상속하도록 구조 변경
