+++ BE/src/test/java/com/issuetracker/service/IssueCommandServiceTest.java
@@ -2,56 +2,24 @@ package com.issuetracker.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.issuetracker.domain.issue.Issue;
-import com.issuetracker.domain.issue.IssueRepository;
 import com.issuetracker.exception.IssueNotFoundException;
 import com.issuetracker.web.dto.reqeust.IssueRequestDTO;
-import org.assertj.core.api.SoftAssertions;
-import org.assertj.core.api.junit.jupiter.InjectSoftAssertions;
-import org.assertj.core.api.junit.jupiter.SoftAssertionsExtension;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.context.WebApplicationContext;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@ExtendWith(SoftAssertionsExtension.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class IssueCommandServiceTest {
-
-    @LocalServerPort
-    private int port;
-
+class IssueCommandServiceTest extends BaseServiceTest {
     @Autowired
     private IssueRepository issueRepository;
-
-    @Autowired
-    private WebApplicationContext context;
-
-    @InjectSoftAssertions
-    SoftAssertions softly;
-
-    private MockMvc mockMvc;
-
-    @BeforeEach
-    public void setup() {
-        mockMvc = MockMvcBuilders
-                .webAppContextSetup(context)
-                .build();
-    }
 
     @Test
     @Transactional
@@ -68,10 +36,15 @@ class IssueCommandServiceTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .header("Authorization", "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJcImlzc3VlLXRyYWNrZXItdGVhbS0wNlwiIiwidXNlcklkIjoxfQ.WCMSnjyZCjZ80aSBN9GCNckS8Q_FkdpWXPWJwsx3kVA")
                 .content(new ObjectMapper().writeValueAsString(issueRequest)))
-                .andExpect(status().isOk());
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.issueId").value(4));
 
-        Issue issue = issueRepository.findById(1L).get();
+        Issue issue = issueRepository.findById(4L).orElseThrow(IssueNotFoundException::new);
         softly.assertThat(issue.getTitle()).isEqualTo(title);
-        softly.assertThat(issue.getFirstComment()).isEqualTo(comment);
+        softly.assertThat(issue.getComments().get(0).getComment()).isEqualTo(comment);
+        softly.assertThat(issue.getAssignees().get(1).getUserName()).isEqualTo("janeljs");
+        softly.assertThat(issue.getLabels().get(0).getName()).isEqualTo("bug");
+        softly.assertThat(issue.isOpen()).isTrue();
+        softly.assertThat(issue.getAuthor().getName()).isEqualTo("Jeong InHo");
     }
 }
